refactor: restore api package
