+++ custom_components/ld2410/api/models.py
