+++ tests/test_init.py
@@ -46,7 +46,7 @@ async def test_exception_handling_for_device_initialization(
     entry.add_to_hass(hass)
 
     with patch(
-        "custom_components.ld2410.ld2410.LD2410Device.__init__",
+        "custom_components.ld2410.api.LD2410Device.__init__",
         side_effect=exception,
     ):
         await hass.config_entries.async_setup(entry.entry_id)
@@ -66,7 +66,7 @@ async def test_setup_entry_without_ble_device(
 
     with (
         patch_async_ble_device_from_address(None),
-        patch("custom_components.ld2410.ld2410.close_stale_connections_by_address"),
+        patch("custom_components.ld2410.api.close_stale_connections_by_address"),
     ):
         await hass.config_entries.async_setup(entry.entry_id)
         await hass.async_block_till_done()
@@ -95,7 +95,7 @@ async def test_coordinator_wait_ready_timeout(
             "custom_components.ld2410.coordinator.asyncio.timeout",
             return_value=timeout_mock,
         ),
-        patch("custom_components.ld2410.ld2410.close_stale_connections_by_address"),
+        patch("custom_components.ld2410.api.close_stale_connections_by_address"),
     ):
         await hass.config_entries.async_setup(entry.entry_id)
         await hass.async_block_till_done()
